# service/swagger-gen/jaxrs-spec-server/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Backend for managing choose-your-own-destiny style stories.",
    "version" : "0.2.0",
    "title" : "StoryTime API",
    "termsOfService" : "http://www.ajegames.com/blog/terms-of-use/",
    "contact" : {
      "name" : "AJE Games Support",
      "url" : "http://www.ajegames.com/blog/support/",
      "email" : "saynotospam@ajegames.com"
    }
  },
  "paths" : {
...
    "/featuredStories" : {
      "get" : {
        "tags" : [ "clubhouse" ],
        "summary" : "Gets a list of featured story IDs.",
        "description" : "Gets a list of featured story IDs. Default list is empty.",
        "operationId" : "getFeaturedStories",
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "Story created",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "admin" ],
        "summary" : "Adds and removes featured stories.",
        "description" : "Adds and removes featured stories.",
        "operationId" : "updateFeaturedStories",
        "parameters" : [ {
          "in" : "body",
          "name" : "toAddRemove",
          "description" : "List of IDs of stories to feature",
          "required" : false,
          "schema" : {
            "type" : "object",
            "properties" : {
              "toAdd" : {
                "type" : "array",
                "items" : {
                  "type" : "string"
                }
              },
              "toRemove" : {
                "type" : "array",
                "items" : {
                  "type" : "string"
                }
              }
            }
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Story created",
            "schema" : {
              "type" : "object",
              "properties" : {
                "storyId" : {
                  "type" : "string"
                },
                "title" : {
                  "type" : "string"
                },
                "authorId" : {
                  "type" : "string"
                },
                "penName" : {
                  "type" : "string"
                },
                "tagLine" : {
                  "type" : "string"
                },
                "about" : {
                  "type" : "string"
                },
                "firstChapter" : {
                  "type" : "string"
                },
                "draft" : {
                  "type" : "boolean"
                },
                "publishedAt" : {
                  "type" : "string",
                  "format" : "date-time"
                },
                "version" : {
                  "type" : "integer",
                  "format" : "int32"
                },
                "genre" : {
                  "type" : "array",
                  "items" : {
                    "type" : "string"
                  }
                },
                "category" : {
                  "type" : "array",
                  "items" : {
                    "type" : "string"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid input",
            "schema" : {
              "type" : "object",
              "properties" : {
                "errorCode" : {
                  "type" : "string"
                },
                "message" : {
                  "type" : "string"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "admin" ],
        "summary" : "Clears list of featured stories.",
        "description" : "Clears list of featured stories.",
        "operationId" : "clearFeaturedStories",
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "Story created"
          }
        }
      }
    },
    "/stories" : {
      "get" : {
        "tags" : [ "storytime" ],
        "summary" : "Gets all of the story summaries",
        "description" : "Gets every story summary. Results might be truncated for paging.",
        "operationId" : "getStorySummaries",
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "Story summaries",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "object",
                "properties" : {
                  "storyId" : {
                    "type" : "string"
                  },
                  "title" : {
                    "type" : "string"
                  },
                  "authorId" : {
                    "type" : "string"
                  },
                  "penName" : {
                    "type" : "string"
                  },
                  "tagLine" : {
                    "type" : "string"
                  },
                  "about" : {
                    "type" : "string"
                  },
                  "firstChapter" : {
                    "type" : "string"
                  },
                  "draft" : {
                    "type" : "boolean"
                  },
                  "publishedAt" : {
                    "type" : "string",
                    "format" : "date-time"
                  },
                  "version" : {
                    "type" : "integer",
                    "format" : "int32"
                  },
                  "genre" : {
                    "type" : "array",
                    "items" : {
                      "type" : "string"
                    }
                  },
                  "category" : {
                    "type" : "array",
                    "items" : {
                      "type" : "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "storytime" ],
        "summary" : "Creates new story with given information.",
        "description" : "Creates a story with the information provided, assigning a unique key.",
        "operationId" : "createStory",
        "parameters" : [ {
          "in" : "body",
          "name" : "story",
          "description" : "Story summary",
          "required" : true,
          "schema" : {
            "type" : "object",
            "properties" : {
              "storyId" : {
                "type" : "string"
              },
              "title" : {
                "type" : "string"
              },
              "penName" : {
                "type" : "string"
              },
              "tagLine" : {
                "type" : "string"
              },
              "about" : {
                "type" : "string"
              },
              "firstChapter" : {
                "type" : "string"
              },
              "draft" : {
                "type" : "boolean"
              },
              "genre" : {
                "type" : "array",
                "items" : {
                  "type" : "string"
                }
              },
              "category" : {
                "type" : "array",
                "items" : {
                  "type" : "string"
                }
              }
            }
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Story created",
            "schema" : {
              "type" : "object",
              "properties" : {
                "storyId" : {
                  "type" : "string"
                },
                "title" : {
                  "type" : "string"
                },
                "authorId" : {
                  "type" : "string"
                },
                "penName" : {
                  "type" : "string"
                },
                "tagLine" : {
                  "type" : "string"
                },
                "about" : {
                  "type" : "string"
                },
                "firstChapter" : {
                  "type" : "string"
                },
                "draft" : {
                  "type" : "boolean"
                },
                "publishedAt" : {
                  "type" : "string",
                  "format" : "date-time"
                },
                "version" : {
                  "type" : "integer",
                  "format" : "int32"
                },
                "genre" : {
                  "type" : "array",
                  "items" : {
                    "type" : "string"
                  }
                },
                "category" : {
                  "type" : "array",
                  "items" : {
                    "type" : "string"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid input",
            "schema" : {
              "type" : "object",
              "properties" : {
                "errorCode" : {
                  "type" : "string"
                },
                "message" : {
                  "type" : "string"
                }
              }
            }
          }
        }
      }
    },
    "/stories/{storyId}" : {
      "get" : {
        "tags" : [ "storytime" ],
        "summary" : "Returns story summary for given key",
        "description" : "Returns the summary of the story indentified by key.",
        "operationId" : "getStorySummary",
        "parameters" : [ {
          "name" : "storyId",
          "in" : "path",
          "description" : "unique key of story to retrieve",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Story found",
            "schema" : {
              "type" : "object",
              "properties" : {
                "storyId" : {
                  "type" : "string"
                },
                "title" : {
                  "type" : "string"
                },
                "authorId" : {
                  "type" : "string"
                },
                "penName" : {
                  "type" : "string"
                },
                "tagLine" : {
                  "type" : "string"
                },
                "about" : {
                  "type" : "string"
                },
                "firstChapter" : {
                  "type" : "string"
                },
                "draft" : {
                  "type" : "boolean"
                },
                "publishedAt" : {
                  "type" : "string",
                  "format" : "date-time"
                },
                "version" : {
                  "type" : "integer",
                  "format" : "int32"
                },
                "genre" : {
                  "type" : "array",
                  "items" : {
                    "type" : "string"
                  }
                },
                "category" : {
                  "type" : "array",
                  "items" : {
                    "type" : "string"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Story not found",
            "schema" : {
              "type" : "object",
              "properties" : {
                "errorCode" : {
                  "type" : "string"
                },
                "message" : {
                  "type" : "string"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "storytime" ],
        "summary" : "Updates a story summary with given information.",
        "description" : "Creates a story with the information provided, assigning a unique key.",
        "operationId" : "updateStorySummary",
        "parameters" : [ {
          "name" : "storyId",
          "in" : "path",
          "description" : "unique key of story to update",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "story",
          "description" : "Story summary",
          "required" : true,
          "schema" : {
            "type" : "object",
            "properties" : {
              "storyId" : {
                "type" : "string"
              },
              "title" : {
                "type" : "string"
              },
              "penName" : {
                "type" : "string"
              },
              "tagLine" : {
                "type" : "string"
              },
              "about" : {
                "type" : "string"
              },
              "firstChapter" : {
                "type" : "string"
              },
              "draft" : {
                "type" : "boolean"
              },
              "genre" : {
                "type" : "array",
                "items" : {
                  "type" : "string"
                }
              },
              "category" : {
                "type" : "array",
                "items" : {
                  "type" : "string"
                }
              }
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Story summary updated",
            "schema" : {
              "type" : "object",
              "properties" : {
                "storyId" : {
                  "type" : "string"
                },
                "title" : {
                  "type" : "string"
                },
                "authorId" : {
                  "type" : "string"
                },
                "penName" : {
                  "type" : "string"
                },
                "tagLine" : {
                  "type" : "string"
                },
                "about" : {
                  "type" : "string"
                },
                "firstChapter" : {
                  "type" : "string"
                },
                "draft" : {
                  "type" : "boolean"
                },
                "publishedAt" : {
                  "type" : "string",
                  "format" : "date-time"
                },
                "version" : {
                  "type" : "integer",
                  "format" : "int32"
                },
                "genre" : {
                  "type" : "array",
                  "items" : {
                    "type" : "string"
                  }
                },
                "category" : {
                  "type" : "array",
                  "items" : {
                    "type" : "string"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid input",
            "schema" : {
              "type" : "object",
              "properties" : {
                "errorCode" : {
                  "type" : "string"
                },
                "message" : {
                  "type" : "string"
                }
              }
            }
          },
          "404" : {
            "description" : "Story not found",
            "schema" : {
              "type" : "object",
              "properties" : {
                "errorCode" : {
                  "type" : "string"
                },
                "message" : {
                  "type" : "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}